fix: moved the folder shared/ from features/ to src/ for respect the architecture
