+++ frontend/src/shared/components/Divider.jsx
