+++ frontend/src/shared/components/GenericTable.jsx
